median filter needs float

+++ tomoprocer.py
@@ -123,7 +123,7 @@ def tomo_prep(cfg, verbose_output=False, write_to_disk=True):
 
     # -- noise reduction
     e = cf.ThreadPoolExecutor(max_workers=multiprocessing.cpu_count() - 1)
-    _jobs = [e.submit(denoise, proj[n,:,:]) for n in range(proj.shape[0])]
+    _jobs = [e.submit(denoise, proj[n,:,:].astype(float)) for n in range(proj.shape[0])]
     # execute
     _proj = [me.result() for me in _jobs]
     # map back
